+++ src/opticalc/integrated.py
@@ -1,20 +1,33 @@
 import logging
 
 import pywincalc
-from py_igsdb_base_data.optical import OpticalStandardMethodResults, OpticalColorResults, \
-    IntegratedSpectralAveragesSummaryValues, OpticalColorFluxResults, OpticalColorResult, ThermalIRResults, \
-    OpticalStandardMethodFluxResults, \
-    IntegratedSpectralAveragesSummaryValuesFactory
+from py_igsdb_base_data.optical import (
+    OpticalStandardMethodResults,
+    OpticalColorResults,
+    IntegratedSpectralAveragesSummaryValues,
+    OpticalColorFluxResults,
+    OpticalColorResult,
+    ThermalIRResults,
+    OpticalStandardMethodFluxResults,
+    IntegratedSpectralAveragesSummaryValuesFactory,
+)
 from py_igsdb_base_data.product import BaseProduct, ProductSubtype
 from py_igsdb_base_data.standard import CalculationStandardMethodTypes
 
 from opticalc.exceptions import SpectralAveragesSummaryCalculationException
-from opticalc.util import convert_to_trichromatic_result, convert_to_lab_result, convert_to_rgb_result, convert_product
+from opticalc.util import (
+    convert_to_trichromatic_result,
+    convert_to_lab_result,
+    convert_to_rgb_result,
+    convert_product,
+)
 
 logger = logging.getLogger(__name__)
 
 
-def calc_optical(glazing_system: pywincalc.GlazingSystem, method_name: str) -> OpticalStandardMethodResults:
+def calc_optical(
+    glazing_system: pywincalc.GlazingSystem, method_name: str
+) -> OpticalStandardMethodResults:
     """
     Uses pywincalc to generate optical information for a given glazing system.
 
@@ -39,9 +52,11 @@ def calc_optical(glazing_system: pywincalc.GlazingSystem, method_name: str) -> O
         results = glazing_system.optical_method_results(method_name)
         system_results = results.system_results
     except Exception as e:
-        err_message = (f"could not create a "
-                       f"pywincalc results object "
-                       f"with method_name: {method_name} : {str(e)}")
+        err_message = (
+            f"could not create a "
+            f"pywincalc results object "
+            f"with method_name: {method_name} : {str(e)}"
+        )
         logger.exception(f"calc_optical() {err_message}")
         raise Exception(err_message) from e
 
@@ -52,31 +67,43 @@ def calc_optical(glazing_system: pywincalc.GlazingSystem, method_name: str) -> O
             direct_diffuse=system_results.front.transmittance.direct_diffuse,
             direct_hemispherical=system_results.front.transmittance.direct_hemispherical,
             diffuse_diffuse=system_results.front.transmittance.diffuse_diffuse,
-            matrix=system_results.front.transmittance.matrix)
+            matrix=system_results.front.transmittance.matrix,
+        )
 
         translated_results.transmittance_back = OpticalStandardMethodFluxResults(
             direct_direct=system_results.back.transmittance.direct_direct,
             direct_diffuse=system_results.back.transmittance.direct_diffuse,
             direct_hemispherical=system_results.back.transmittance.direct_hemispherical,
             diffuse_diffuse=system_results.back.transmittance.diffuse_diffuse,
-            matrix=system_results.back.transmittance.matrix)
+            matrix=system_results.back.transmittance.matrix,
+        )
 
         translated_results.reflectance_front = OpticalStandardMethodFluxResults(
             direct_direct=system_results.front.reflectance.direct_direct,
             direct_diffuse=system_results.front.reflectance.direct_diffuse,
             direct_hemispherical=system_results.front.reflectance.direct_hemispherical,
             diffuse_diffuse=system_results.front.reflectance.diffuse_diffuse,
-            matrix=system_results.front.reflectance.matrix)
+            matrix=system_results.front.reflectance.matrix,
+        )
 
         translated_results.reflectance_back = OpticalStandardMethodFluxResults(
             direct_direct=system_results.back.reflectance.direct_direct,
             direct_diffuse=system_results.back.reflectance.direct_diffuse,
             direct_hemispherical=system_results.back.reflectance.direct_hemispherical,
             diffuse_diffuse=system_results.back.reflectance.diffuse_diffuse,
-            matrix=system_results.back.reflectance.matrix)
+            matrix=system_results.back.reflectance.matrix,
+        )
 
-        translated_results.absorptance_front = 1 - system_results.front.transmittance.direct_hemispherical - system_results.front.reflectance.direct_hemispherical
-        translated_results.absorptance_back = 1 - system_results.back.transmittance.direct_hemispherical - system_results.back.reflectance.direct_hemispherical
+        translated_results.absorptance_front = (
+            1
+            - system_results.front.transmittance.direct_hemispherical
+            - system_results.front.reflectance.direct_hemispherical
+        )
+        translated_results.absorptance_back = (
+            1
+            - system_results.back.transmittance.direct_hemispherical
+            - system_results.back.reflectance.direct_hemispherical
+        )
 
     except Exception as e:
         logger.exception(f"calc_optical() call failed for method {method_name}")
@@ -110,175 +137,283 @@ def calc_color(glazing_system: pywincalc.GlazingSystem) -> OpticalColorResults:
         results = results.system_results
 
         direct_direct_front_transmittance_trichromatic = convert_to_trichromatic_result(
-            results.front.transmittance.direct_direct.trichromatic)
-        direct_direct_front_transmittance_lab = convert_to_lab_result(results.front.transmittance.direct_direct.lab)
-        direct_direct_front_transmittance_rgb = convert_to_rgb_result(results.front.transmittance.direct_direct.rgb)
+            results.front.transmittance.direct_direct.trichromatic
+        )
+        direct_direct_front_transmittance_lab = convert_to_lab_result(
+            results.front.transmittance.direct_direct.lab
+        )
+        direct_direct_front_transmittance_rgb = convert_to_rgb_result(
+            results.front.transmittance.direct_direct.rgb
+        )
 
-        direct_diffuse_front_transmittance_trichromatic = convert_to_trichromatic_result(
-            results.front.transmittance.direct_diffuse.trichromatic)
-        direct_diffuse_front_transmittance_lab = convert_to_lab_result(results.front.transmittance.direct_diffuse.lab)
-        direct_diffuse_front_transmittance_rgb = convert_to_rgb_result(results.front.transmittance.direct_diffuse.rgb)
+        direct_diffuse_front_transmittance_trichromatic = (
+            convert_to_trichromatic_result(
+                results.front.transmittance.direct_diffuse.trichromatic
+            )
+        )
+        direct_diffuse_front_transmittance_lab = convert_to_lab_result(
+            results.front.transmittance.direct_diffuse.lab
+        )
+        direct_diffuse_front_transmittance_rgb = convert_to_rgb_result(
+            results.front.transmittance.direct_diffuse.rgb
+        )
 
-        diffuse_diffuse_front_transmittance_trichromatic = convert_to_trichromatic_result(
-            results.front.transmittance.diffuse_diffuse.trichromatic)
-        diffuse_diffuse_front_transmittance_lab = convert_to_lab_result(results.front.transmittance.diffuse_diffuse.lab)
-        diffuse_diffuse_front_transmittance_rgb = convert_to_rgb_result(results.front.transmittance.diffuse_diffuse.rgb)
+        diffuse_diffuse_front_transmittance_trichromatic = (
+            convert_to_trichromatic_result(
+                results.front.transmittance.diffuse_diffuse.trichromatic
+            )
+        )
+        diffuse_diffuse_front_transmittance_lab = convert_to_lab_result(
+            results.front.transmittance.diffuse_diffuse.lab
+        )
+        diffuse_diffuse_front_transmittance_rgb = convert_to_rgb_result(
+            results.front.transmittance.diffuse_diffuse.rgb
+        )
 
-        direct_hemispherical_front_transmittance_trichromatic = convert_to_trichromatic_result(
-            results.front.transmittance.direct_hemispherical.trichromatic)
+        direct_hemispherical_front_transmittance_trichromatic = (
+            convert_to_trichromatic_result(
+                results.front.transmittance.direct_hemispherical.trichromatic
+            )
+        )
         direct_hemispherical_front_transmittance_lab = convert_to_lab_result(
-            results.front.transmittance.direct_hemispherical.lab)
+            results.front.transmittance.direct_hemispherical.lab
+        )
         direct_hemispherical_front_transmittance_rgb = convert_to_rgb_result(
-            results.front.transmittance.direct_hemispherical.rgb)
+            results.front.transmittance.direct_hemispherical.rgb
+        )
 
         translated_results.transmittance_front = OpticalColorFluxResults(
             direct_direct=OpticalColorResult(
                 trichromatic=direct_direct_front_transmittance_trichromatic,
                 lab=direct_direct_front_transmittance_lab,
-                rgb=direct_direct_front_transmittance_rgb),
+                rgb=direct_direct_front_transmittance_rgb,
+            ),
             direct_diffuse=OpticalColorResult(
                 trichromatic=direct_diffuse_front_transmittance_trichromatic,
                 lab=direct_diffuse_front_transmittance_lab,
-                rgb=direct_diffuse_front_transmittance_rgb),
+                rgb=direct_diffuse_front_transmittance_rgb,
+            ),
             direct_hemispherical=OpticalColorResult(
                 trichromatic=direct_hemispherical_front_transmittance_trichromatic,
                 lab=direct_hemispherical_front_transmittance_lab,
-                rgb=direct_hemispherical_front_transmittance_rgb),
+                rgb=direct_hemispherical_front_transmittance_rgb,
+            ),
             diffuse_diffuse=OpticalColorResult(
                 trichromatic=diffuse_diffuse_front_transmittance_trichromatic,
                 lab=diffuse_diffuse_front_transmittance_lab,
-                rgb=diffuse_diffuse_front_transmittance_rgb))
+                rgb=diffuse_diffuse_front_transmittance_rgb,
+            ),
+        )
 
         direct_direct_front_reflectance_trichromatic = convert_to_trichromatic_result(
-            results.front.reflectance.direct_direct.trichromatic)
-        direct_direct_front_reflectance_lab = convert_to_lab_result(results.front.reflectance.direct_direct.lab)
-        direct_direct_front_reflectance_rgb = convert_to_rgb_result(results.front.reflectance.direct_direct.rgb)
+            results.front.reflectance.direct_direct.trichromatic
+        )
+        direct_direct_front_reflectance_lab = convert_to_lab_result(
+            results.front.reflectance.direct_direct.lab
+        )
+        direct_direct_front_reflectance_rgb = convert_to_rgb_result(
+            results.front.reflectance.direct_direct.rgb
+        )
 
         direct_diffuse_front_reflectance_trichromatic = convert_to_trichromatic_result(
-            results.front.reflectance.direct_diffuse.trichromatic)
-        direct_diffuse_front_reflectance_lab = convert_to_lab_result(results.front.reflectance.direct_diffuse.lab)
-        direct_diffuse_front_reflectance_rgb = convert_to_rgb_result(results.front.reflectance.direct_diffuse.rgb)
+            results.front.reflectance.direct_diffuse.trichromatic
+        )
+        direct_diffuse_front_reflectance_lab = convert_to_lab_result(
+            results.front.reflectance.direct_diffuse.lab
+        )
+        direct_diffuse_front_reflectance_rgb = convert_to_rgb_result(
+            results.front.reflectance.direct_diffuse.rgb
+        )
 
-        direct_hemispherical_front_reflectance_trichromatic = convert_to_trichromatic_result(
-            results.front.reflectance.direct_hemispherical.trichromatic)
+        direct_hemispherical_front_reflectance_trichromatic = (
+            convert_to_trichromatic_result(
+                results.front.reflectance.direct_hemispherical.trichromatic
+            )
+        )
         direct_hemispherical_front_reflectance_lab = convert_to_lab_result(
-            results.front.reflectance.direct_hemispherical.lab)
+            results.front.reflectance.direct_hemispherical.lab
+        )
         direct_hemispherical_front_reflectance_rgb = convert_to_rgb_result(
-            results.front.reflectance.direct_hemispherical.rgb)
+            results.front.reflectance.direct_hemispherical.rgb
+        )
 
         diffuse_diffuse_front_reflectance_trichromatic = convert_to_trichromatic_result(
-            results.front.reflectance.diffuse_diffuse.trichromatic)
-        diffuse_diffuse_front_reflectance_lab = convert_to_lab_result(results.front.reflectance.diffuse_diffuse.lab)
-        diffuse_diffuse_front_reflectance_rgb = convert_to_rgb_result(results.front.reflectance.diffuse_diffuse.rgb)
+            results.front.reflectance.diffuse_diffuse.trichromatic
+        )
+        diffuse_diffuse_front_reflectance_lab = convert_to_lab_result(
+            results.front.reflectance.diffuse_diffuse.lab
+        )
+        diffuse_diffuse_front_reflectance_rgb = convert_to_rgb_result(
+            results.front.reflectance.diffuse_diffuse.rgb
+        )
 
         translated_results.reflectance_front = OpticalColorFluxResults(
             direct_direct=OpticalColorResult(
                 trichromatic=direct_direct_front_reflectance_trichromatic,
                 lab=direct_direct_front_reflectance_lab,
-                rgb=direct_direct_front_reflectance_rgb),
+                rgb=direct_direct_front_reflectance_rgb,
+            ),
             direct_diffuse=OpticalColorResult(
                 trichromatic=direct_diffuse_front_reflectance_trichromatic,
                 lab=direct_diffuse_front_reflectance_lab,
-                rgb=direct_diffuse_front_reflectance_rgb),
+                rgb=direct_diffuse_front_reflectance_rgb,
+            ),
             direct_hemispherical=OpticalColorResult(
                 trichromatic=direct_hemispherical_front_reflectance_trichromatic,
                 lab=direct_hemispherical_front_reflectance_lab,
-                rgb=direct_hemispherical_front_reflectance_rgb),
+                rgb=direct_hemispherical_front_reflectance_rgb,
+            ),
             diffuse_diffuse=OpticalColorResult(
                 trichromatic=diffuse_diffuse_front_reflectance_trichromatic,
                 lab=diffuse_diffuse_front_reflectance_lab,
-                rgb=diffuse_diffuse_front_reflectance_rgb))
+                rgb=diffuse_diffuse_front_reflectance_rgb,
+            ),
+        )
 
         direct_direct_back_transmittance_trichromatic = convert_to_trichromatic_result(
-            results.back.transmittance.direct_direct.trichromatic)
-        direct_direct_back_transmittance_lab = convert_to_lab_result(results.back.transmittance.direct_direct.lab)
-        direct_direct_back_transmittance_rgb = convert_to_rgb_result(results.back.transmittance.direct_direct.rgb)
+            results.back.transmittance.direct_direct.trichromatic
+        )
+        direct_direct_back_transmittance_lab = convert_to_lab_result(
+            results.back.transmittance.direct_direct.lab
+        )
+        direct_direct_back_transmittance_rgb = convert_to_rgb_result(
+            results.back.transmittance.direct_direct.rgb
+        )
 
         direct_diffuse_back_transmittance_trichromatic = convert_to_trichromatic_result(
-            results.back.transmittance.direct_diffuse.trichromatic)
-        direct_diffuse_back_transmittance_lab = convert_to_lab_result(results.back.transmittance.direct_diffuse.lab)
-        direct_diffuse_back_transmittance_rgb = convert_to_rgb_result(results.back.transmittance.direct_diffuse.rgb)
+            results.back.transmittance.direct_diffuse.trichromatic
+        )
+        direct_diffuse_back_transmittance_lab = convert_to_lab_result(
+            results.back.transmittance.direct_diffuse.lab
+        )
+        direct_diffuse_back_transmittance_rgb = convert_to_rgb_result(
+            results.back.transmittance.direct_diffuse.rgb
+        )
 
-        direct_hemispherical_back_transmittance_trichromatic = convert_to_trichromatic_result(
-            results.back.transmittance.direct_hemispherical.trichromatic)
+        direct_hemispherical_back_transmittance_trichromatic = (
+            convert_to_trichromatic_result(
+                results.back.transmittance.direct_hemispherical.trichromatic
+            )
+        )
         direct_hemispherical_back_transmittance_lab = convert_to_lab_result(
-            results.back.transmittance.direct_hemispherical.lab)
+            results.back.transmittance.direct_hemispherical.lab
+        )
         direct_hemispherical_back_transmittance_rgb = convert_to_rgb_result(
-            results.back.transmittance.direct_hemispherical.rgb)
+            results.back.transmittance.direct_hemispherical.rgb
+        )
 
-        diffuse_diffuse_back_transmittance_trichromatic = convert_to_trichromatic_result(
-            results.back.transmittance.diffuse_diffuse.trichromatic)
-        diffuse_diffuse_back_transmittance_lab = convert_to_lab_result(results.back.transmittance.diffuse_diffuse.lab)
-        diffuse_diffuse_back_transmittance_rgb = convert_to_rgb_result(results.back.transmittance.diffuse_diffuse.rgb)
+        diffuse_diffuse_back_transmittance_trichromatic = (
+            convert_to_trichromatic_result(
+                results.back.transmittance.diffuse_diffuse.trichromatic
+            )
+        )
+        diffuse_diffuse_back_transmittance_lab = convert_to_lab_result(
+            results.back.transmittance.diffuse_diffuse.lab
+        )
+        diffuse_diffuse_back_transmittance_rgb = convert_to_rgb_result(
+            results.back.transmittance.diffuse_diffuse.rgb
+        )
 
         translated_results.transmittance_back = OpticalColorFluxResults(
             direct_direct=OpticalColorResult(
                 trichromatic=direct_direct_back_transmittance_trichromatic,
                 lab=direct_direct_back_transmittance_lab,
-                rgb=direct_direct_back_transmittance_rgb),
+                rgb=direct_direct_back_transmittance_rgb,
+            ),
             direct_diffuse=OpticalColorResult(
                 trichromatic=direct_diffuse_back_transmittance_trichromatic,
                 lab=direct_diffuse_back_transmittance_lab,
-                rgb=direct_diffuse_back_transmittance_rgb),
+                rgb=direct_diffuse_back_transmittance_rgb,
+            ),
             direct_hemispherical=OpticalColorResult(
                 trichromatic=direct_hemispherical_back_transmittance_trichromatic,
                 lab=direct_hemispherical_back_transmittance_lab,
-                rgb=direct_hemispherical_back_transmittance_rgb),
+                rgb=direct_hemispherical_back_transmittance_rgb,
+            ),
             diffuse_diffuse=OpticalColorResult(
                 trichromatic=diffuse_diffuse_back_transmittance_trichromatic,
                 lab=diffuse_diffuse_back_transmittance_lab,
-                rgb=diffuse_diffuse_back_transmittance_rgb))
+                rgb=diffuse_diffuse_back_transmittance_rgb,
+            ),
+        )
 
         direct_direct_back_reflectance_trichromatic = convert_to_trichromatic_result(
-            results.back.reflectance.direct_direct.trichromatic)
-        direct_direct_back_reflectance_lab = convert_to_lab_result(results.back.reflectance.direct_direct.lab)
-        direct_direct_back_reflectance_rgb = convert_to_rgb_result(results.back.reflectance.direct_direct.rgb)
+            results.back.reflectance.direct_direct.trichromatic
+        )
+        direct_direct_back_reflectance_lab = convert_to_lab_result(
+            results.back.reflectance.direct_direct.lab
+        )
+        direct_direct_back_reflectance_rgb = convert_to_rgb_result(
+            results.back.reflectance.direct_direct.rgb
+        )
 
         direct_diffuse_back_reflectance_trichromatic = convert_to_trichromatic_result(
-            results.back.reflectance.direct_diffuse.trichromatic)
-        direct_diffuse_back_reflectance_lab = convert_to_lab_result(results.back.reflectance.direct_diffuse.lab)
-        direct_diffuse_back_reflectance_rgb = convert_to_rgb_result(results.back.reflectance.direct_diffuse.rgb)
+            results.back.reflectance.direct_diffuse.trichromatic
+        )
+        direct_diffuse_back_reflectance_lab = convert_to_lab_result(
+            results.back.reflectance.direct_diffuse.lab
+        )
+        direct_diffuse_back_reflectance_rgb = convert_to_rgb_result(
+            results.back.reflectance.direct_diffuse.rgb
+        )
 
-        direct_hemispherical_back_reflectance_trichromatic = convert_to_trichromatic_result(
-            results.back.reflectance.direct_hemispherical.trichromatic)
+        direct_hemispherical_back_reflectance_trichromatic = (
+            convert_to_trichromatic_result(
+                results.back.reflectance.direct_hemispherical.trichromatic
+            )
+        )
         direct_hemispherical_back_reflectance_lab = convert_to_lab_result(
-            results.back.reflectance.direct_hemispherical.lab)
+            results.back.reflectance.direct_hemispherical.lab
+        )
         direct_hemispherical_back_reflectance_rgb = convert_to_rgb_result(
-            results.back.reflectance.direct_hemispherical.rgb)
+            results.back.reflectance.direct_hemispherical.rgb
+        )
 
         diffuse_diffuse_back_reflectance_trichromatic = convert_to_trichromatic_result(
-            results.back.reflectance.diffuse_diffuse.trichromatic)
-        diffuse_diffuse_back_reflectance_lab = convert_to_lab_result(results.back.reflectance.diffuse_diffuse.lab)
-        diffuse_diffuse_back_reflectance_rgb = convert_to_rgb_result(results.back.reflectance.diffuse_diffuse.rgb)
+            results.back.reflectance.diffuse_diffuse.trichromatic
+        )
+        diffuse_diffuse_back_reflectance_lab = convert_to_lab_result(
+            results.back.reflectance.diffuse_diffuse.lab
+        )
+        diffuse_diffuse_back_reflectance_rgb = convert_to_rgb_result(
+            results.back.reflectance.diffuse_diffuse.rgb
+        )
 
         translated_results.reflectance_back = OpticalColorFluxResults(
             direct_direct=OpticalColorResult(
                 trichromatic=direct_direct_back_reflectance_trichromatic,
                 lab=direct_direct_back_reflectance_lab,
-                rgb=direct_direct_back_reflectance_rgb),
+                rgb=direct_direct_back_reflectance_rgb,
+            ),
             direct_diffuse=OpticalColorResult(
                 trichromatic=direct_diffuse_back_reflectance_trichromatic,
                 lab=direct_diffuse_back_reflectance_lab,
-                rgb=direct_diffuse_back_reflectance_rgb),
+                rgb=direct_diffuse_back_reflectance_rgb,
+            ),
             direct_hemispherical=OpticalColorResult(
                 trichromatic=direct_hemispherical_back_reflectance_trichromatic,
                 lab=direct_hemispherical_back_reflectance_lab,
-                rgb=direct_hemispherical_back_reflectance_rgb),
+                rgb=direct_hemispherical_back_reflectance_rgb,
+            ),
             diffuse_diffuse=OpticalColorResult(
                 trichromatic=diffuse_diffuse_back_reflectance_trichromatic,
                 lab=diffuse_diffuse_back_reflectance_lab,
-                rgb=diffuse_diffuse_back_reflectance_rgb))
+                rgb=diffuse_diffuse_back_reflectance_rgb,
+            ),
+        )
 
     except Exception as e:
-        logger.exception(f"calc_color() call failed")
+        logger.exception("calc_color() call failed")
         raise e
 
     return translated_results
 
 
-def calc_thermal_ir_results(optical_standard: pywincalc.OpticalStandard,
-                            pywincalc_layer: pywincalc.ProductDataOpticalAndThermal,
-                            ignore_emissivity: bool = False) -> ThermalIRResults:
+def calc_thermal_ir_results(
+    optical_standard: pywincalc.OpticalStandard,
+    pywincalc_layer: pywincalc.ProductDataOpticalAndThermal,
+    ignore_emissivity: bool = False,
+) -> ThermalIRResults:
     """
     Uses pywincalc to generate thermal IR information for a given layer and standard.
 
@@ -304,24 +439,35 @@ def calc_thermal_ir_results(optical_standard: pywincalc.OpticalStandard,
     try:
         translated_results = ThermalIRResults()
         pywincalc_results = pywincalc.calc_thermal_ir(optical_standard, pywincalc_layer)
-        translated_results.transmittance_front_diffuse_diffuse = pywincalc_results.transmittance_front_diffuse_diffuse
-        translated_results.transmittance_back_diffuse_diffuse = pywincalc_results.transmittance_back_diffuse_diffuse
+        translated_results.transmittance_front_diffuse_diffuse = (
+            pywincalc_results.transmittance_front_diffuse_diffuse
+        )
+        translated_results.transmittance_back_diffuse_diffuse = (
+            pywincalc_results.transmittance_back_diffuse_diffuse
+        )
         if ignore_emissivity:
-            logger.info("Ignoring emissivity_front_hemispheric and emissivity_back_hemispheric from ThermalIRResults()")
+            logger.info(
+                "Ignoring emissivity_front_hemispheric and emissivity_back_hemispheric from ThermalIRResults()"
+            )
         else:
-            translated_results.emissivity_front_hemispheric = pywincalc_results.emissivity_front_hemispheric
-            translated_results.emissivity_back_hemispheric = pywincalc_results.emissivity_back_hemispheric
+            translated_results.emissivity_front_hemispheric = (
+                pywincalc_results.emissivity_front_hemispheric
+            )
+            translated_results.emissivity_back_hemispheric = (
+                pywincalc_results.emissivity_back_hemispheric
+            )
     except Exception as e:
-        logger.exception(f"calc_thermal_ir_results() call failed")
+        logger.exception("calc_thermal_ir_results() call failed")
         raise e
 
     return translated_results
 
 
-def generate_integrated_spectral_averages_summary(product: BaseProduct,
-                                                  optical_standard: pywincalc.OpticalStandard,
-                                                  use_diffuse_as_specular: bool = False) \
-        -> IntegratedSpectralAveragesSummaryValues:
+def generate_integrated_spectral_averages_summary(
+    product: BaseProduct,
+    optical_standard: pywincalc.OpticalStandard,
+    use_diffuse_as_specular: bool = False,
+) -> IntegratedSpectralAveragesSummaryValues:
     """
     Uses pywincalc to generate an integrated spectral averages summary for a given product
     and standard. Returns information in a populated instance of the IntegratedSpectralAveragesSummaryValues dataclass.
@@ -360,14 +506,18 @@ def generate_integrated_spectral_averages_summary(product: BaseProduct,
         raise ValueError("optical_standard is None")
 
     # Create an empty summary_results instance. We'll populate it below and then return it.
-    summary_results: IntegratedSpectralAveragesSummaryValues = IntegratedSpectralAveragesSummaryValuesFactory.create()
+    summary_results: IntegratedSpectralAveragesSummaryValues = (
+        IntegratedSpectralAveragesSummaryValuesFactory.create()
+    )
 
     # Convert product from a BaseData object into a format that pywincalc understands.
-    pywincalc_layer: pywincalc.ProductDataOpticalAndThermal = convert_product(product,
-                                                                              use_diffuse_as_specular=use_diffuse_as_specular)
+    pywincalc_layer: pywincalc.ProductDataOpticalAndThermal = convert_product(
+        product, use_diffuse_as_specular=use_diffuse_as_specular
+    )
 
-    glazing_system: pywincalc.GlazingSystem = pywincalc.GlazingSystem(optical_standard=optical_standard,
-                                                                      solid_layers=[pywincalc_layer])
+    glazing_system: pywincalc.GlazingSystem = pywincalc.GlazingSystem(
+        optical_standard=optical_standard, solid_layers=[pywincalc_layer]
+    )
 
     # Iterate through optical methods and add the calculated results to the
     # summary_results instance. Note that we skip a method if
@@ -377,11 +527,9 @@ def generate_integrated_spectral_averages_summary(product: BaseProduct,
         CalculationStandardMethodTypes.SOLAR.name,
         CalculationStandardMethodTypes.TDW.name,
         CalculationStandardMethodTypes.TKR.name,
-        CalculationStandardMethodTypes.TUV.name
-
+        CalculationStandardMethodTypes.TUV.name,
         # NOTE:
         #   We do not include THERMAL_IR here as we do it further below.
-
         # NOTE:
         #   We skip SPF during optical calcs because of the following reasons
         #   provided by JJ and SC:
@@ -391,30 +539,35 @@ def generate_integrated_spectral_averages_summary(product: BaseProduct,
         #           transmittance that prevents it from being directly calculated with the function used for all
         #           other spectral average metrics. (IIRC you get a T_spf using the spectral averaging routine and
         #           then report SPF as 1/T_spf).
-
         # CalculationStandardMethodTypes.SPF.name
     ]
     for method_name in optical_methods:
         if method_name in optical_standard.methods:
             try:
-                results: OpticalStandardMethodResults = calc_optical(glazing_system, method_name)
+                results: OpticalStandardMethodResults = calc_optical(
+                    glazing_system, method_name
+                )
                 setattr(summary_results, method_name.lower(), results)
             except Exception as e:
                 error_msg = f"calc_optical() call failed for method {method_name} : {e}"
                 logger.error(f"OptiCalc : {error_msg} ")
                 raise SpectralAveragesSummaryCalculationException(error_msg) from e
         else:
-            logger.warning(f"generate_integrated_spectral_averages_summary() skipping method {method_name} as its not "
-                           f"present in the methods for optical standard {optical_standard} "
-                           f"( methods : {optical_standard.methods} )")
+            logger.warning(
+                f"generate_integrated_spectral_averages_summary() skipping method {method_name} as its not "
+                f"present in the methods for optical standard {optical_standard} "
+                f"( methods : {optical_standard.methods} )"
+            )
 
     # Add color results to summary_results
     try:
         summary_results.color = calc_color(glazing_system)
     except Exception as e:
-        error_msg = f"calc_color() call failed for product: {product} " \
-                    f"optical_standard : {optical_standard} " \
-                    f"glazing_system {glazing_system}"
+        error_msg = (
+            f"calc_color() call failed for product: {product} "
+            f"optical_standard : {optical_standard} "
+            f"glazing_system {glazing_system}"
+        )
         logger.error(f"OptiCalc : {error_msg}  error : {e}")
         raise SpectralAveragesSummaryCalculationException(error_msg) from e
 
@@ -428,13 +581,15 @@ def generate_integrated_spectral_averages_summary(product: BaseProduct,
     # We only want to refer to header value emissivity for products of this subtype.
     ignore_emissivity = product.subtype == ProductSubtype.SHADE_MATERIAL.name
     try:
-        summary_results.thermal_ir = calc_thermal_ir_results(optical_standard,
-                                                             pywincalc_layer,
-                                                             ignore_emissivity=ignore_emissivity)
+        summary_results.thermal_ir = calc_thermal_ir_results(
+            optical_standard, pywincalc_layer, ignore_emissivity=ignore_emissivity
+        )
     except Exception as e:
-        error_msg = f"calc_thermal_ir_results() call failed for " \
-                    f"layer: {pywincalc_layer} " \
-                    f"optical_standard: {optical_standard} "
+        error_msg = (
+            f"calc_thermal_ir_results() call failed for "
+            f"layer: {pywincalc_layer} "
+            f"optical_standard: {optical_standard} "
+        )
         logger.error(f"OptiCalc : {error_msg} : {e}")
         raise SpectralAveragesSummaryCalculationException(error_msg) from e